+++ src/tables/cmap/format0.rs
@@ -5,12 +5,16 @@ use crate::parser::{Stream, NumFrom};
 pub fn parse(data: &[u8], code_point: u32) -> Option<u16> {
     let mut s = Stream::new(data);
     s.skip::<u16>(); // format
-    let length: u16 = s.read()?;
+    s.skip::<u16>(); // length
     s.skip::<u16>(); // language
 
-    if code_point < u32::from(length) {
-        s.advance(usize::num_from(code_point));
-        Some(u16::from(s.read::<u8>()?))
+    s.advance(usize::num_from(code_point));
+    let glyph_id: u8 = s.read()?;
+
+    // Make sure that the glyph is not zero, the array always has length 256,
+    // but some codepoints may be mapped to zero.
+    if glyph_id != 0 {
+        Some(u16::from(glyph_id))
     } else {
         None
     }
@@ -39,7 +43,7 @@ pub fn codepoints(data: &[u8], mut f: impl FnMut(u32)) -> Option<()> {
 
 #[cfg(test)]
 mod format0_tests {
-    use super::codepoints;
+    use super::{parse, codepoints};
 
     #[test]
     fn maps_not_all_256_codepoints() {
@@ -53,6 +57,10 @@ mod format0_tests {
         data.extend(std::iter::repeat(0).take(256));
         data[6 + 0x40] = 100;
 
+        assert_eq!(parse(&data, 0), None);
+        assert_eq!(parse(&data, 0x40), Some(100));
+        assert_eq!(parse(&data, 100), None);
+      
         let mut vec = vec![];
         codepoints(&data, |c| vec.push(c));
         assert_eq!(vec, [0x40]);